requirements.txt
streamlit
weaviate-client
